refactor auth functionalities
