+++ server/config/variables.js
@@ -1,0 +1,6 @@
+import dotenv from 'dotenv';
+
+dotenv.config();
+
+export const frontendUrl = process.env.FRONTEND_URL;
+export const backendUrl = process.env.BACKEND_URL;
