+++ server/middleware/index.js
@@ -1,0 +1,7 @@
+import UserValidation from './user-validation';
+import Authorization from './authorization';
+
+export {
+  UserValidation,
+  Authorization
+};
