+++ server/routes/auth.routes.js
@@ -1,26 +1,27 @@
-import express, { Router } from 'express';
+import { Router } from 'express';
 import passport from 'passport';
-import { UserController } from '../controllers';
-import {
-  tryCatch,
+import { UserController, AuthController } from '../controllers';
+import { UserValidation } from '../middleware';
+import tryCatch from '../utils/trycatch.util';
+
+const {
   userSignup,
   userExist,
   isSigninFieldEmpty,
   checkPassword,
   userNotExist,
   checkUsername
-} from '../utils';
+} = UserValidation;
 
 const { socialRedirect } = UserController;
 
 const router = new Router();
 
-router.post('/signup', [checkUsername, userSignup, userExist], tryCatch(UserController.create));
+router.post('/signup', [checkUsername, userSignup, userExist], tryCatch(AuthController.signup));
+router.post('/login', [isSigninFieldEmpty], tryCatch(AuthController.signinUser));
 
-router.post('/login', isSigninFieldEmpty, tryCatch(UserController.signinUser));
-router.post('/forgotpassword', userNotExist, tryCatch(UserController.forgotPassword));
-
-router.put('/resetpassword', checkPassword, tryCatch(UserController.resetPassword));
+router.post('/forgotpassword', [userNotExist], tryCatch(UserController.forgotPassword));
+router.put('/resetpassword', [checkPassword], tryCatch(UserController.resetPassword));
 
 router.get('/facebook', passport.authenticate('facebook', { scope: ['email'] }));
 router.get('/facebook/callback', passport.authenticate('facebook', { session: false }), socialRedirect);
