+++ server/routes/profile.routes.js
@@ -1,12 +1,7 @@
-import express, { Router } from 'express';
+import { Router } from 'express';
 import { ProfileController } from '../controllers';
-import { signInAuth } from '../utils/users/permissions.util';
-import {
-  tryCatch,
-  UuidValidator,
-  validateProfile,
-  imageUpload,
-} from '../utils';
+import { tryCatch, UuidValidator, imageUpload } from '../utils';
+import { UserValidation, Authorization } from '../middleware';
 
 const {
   followUser,
@@ -18,14 +13,16 @@ const {
   getFollowee
 } = ProfileController;
 const { validId } = UuidValidator;
+const { validateProfile } = UserValidation;
+const { signInAuth } = Authorization;
 const router = new Router();
 
 router.post('/:username/follow', [signInAuth], tryCatch(followUser));
 router.delete('/:username/follow', [signInAuth], tryCatch(unfollowUser));
-router.get('/followers', signInAuth, tryCatch(getFollowers));
-router.get('/followee', signInAuth, tryCatch(getFollowee));
-router.get('/notifications', signInAuth, tryCatch(getNotifications));
-router.get('/:id', signInAuth, validId, tryCatch(getProfileById));
-router.patch('/', signInAuth, imageUpload, validateProfile, tryCatch(updateProfile));
+router.get('/followers', [signInAuth], tryCatch(getFollowers));
+router.get('/followee', [signInAuth], tryCatch(getFollowee));
+router.get('/notifications', [signInAuth], tryCatch(getNotifications));
+router.get('/:id', [signInAuth, validId], tryCatch(getProfileById));
+router.patch('/', [signInAuth, imageUpload, validateProfile], tryCatch(updateProfile));
 
 export default router;
