+++ server/routes/users.routes.js
@@ -1,23 +1,17 @@
 import express, { Router } from 'express';
-import {
-  UserController,
-  BookmarkController,
-  ProfileController
-} from '../controllers';
-import { signInAuth, superAdminAuth } from '../utils/users/permissions.util';
-import {
-  tryCatch,
-  UuidValidator,
-  validateRole,
-} from '../utils';
+import { UserController, BookmarkController } from '../controllers';
+import { UserValidation, Authorization } from '../middleware';
+import { tryCatch, UuidValidator } from '../utils';
 
 const { updateUserRole, getUsers } = UserController;
 const { getUserBookmark } = BookmarkController;
 const { validId } = UuidValidator;
+const { validateRole } = UserValidation;
+const { signInAuth, superAdminAuth } = Authorization;
 const router = new Router();
 
-router.patch('/role/:id', signInAuth, superAdminAuth, validId, validateRole, tryCatch(updateUserRole));
-router.get('/bookmark', signInAuth, tryCatch(getUserBookmark));
-router.get('/', signInAuth, tryCatch(getUsers));
+router.patch('/role/:id', [signInAuth, superAdminAuth, validId, validateRole], tryCatch(updateUserRole));
+router.get('/bookmark', [signInAuth], tryCatch(getUserBookmark));
+router.get('/', [signInAuth], tryCatch(getUsers));
 
 export default router;
