+++ server/test/signup.spec.js
@@ -35,69 +35,69 @@ describe('Signup routes', () => {
       expect(res.body).to.have.property('message').eql({ lastname: 'lastname is required.' });
     });
 
-    it('should return 422 if email not sent', async () => {
+    it('should return 400 if email not sent', async () => {
       const res = await chai.request(app)
         .post('/api/v1/auth/signup')
         .send(userMockData.invalidData[2]);
 
-      expect(res).to.have.status(422);
+      expect(res).to.have.status(400);
       expect(res.body).to.have.property('message').eql({ email: 'Email value  is invalid' });
     });
 
-    it('should return 422 if password not sent', async () => {
+    it('should return 400 if password not sent', async () => {
       const res = await chai.request(app)
         .post('/api/v1/auth/signup')
         .send(userMockData.invalidData[3]);
 
-      expect(res).to.have.status(422);
+      expect(res).to.have.status(400);
       expect(res.body).to.have.property('message').eql({
         passwordLength: 'Password should be a minimum of 8 characters',
         passwordType: 'Password should be alphanumeric'
       });
     });
 
-    it('should return 422 if user already exist', async () => {
+    it('should return 409 if user already exist', async () => {
       const res = await chai.request(app)
         .post('/api/v1/auth/signup')
         .send(userMockData.invalidData[4]);
 
-      expect(res).to.have.status(422);
+      expect(res).to.have.status(409);
       expect(res.body).to.have.property('message').eql('User with email: d.ogundimy@andela.com already exist!');
     });
 
-    it('should return 422 if email is invalid', async () => {
+    it('should return 400 if email is invalid', async () => {
       const res = await chai.request(app)
         .post('/api/v1/auth/signup')
         .send(userMockData.invalidData[5]);
 
-      expect(res).to.have.status(422);
+      expect(res).to.have.status(400);
       expect(res.body).to.have.property('message').eql({ email: 'Email value d.ogundimyandela.com is invalid' });
     });
 
-    it('should return 422 if password length does not meet requirements', async () => {
+    it('should return 400 if password length does not meet requirements', async () => {
       const res = await chai.request(app)
         .post('/api/v1/auth/signup')
         .send(userMockData.invalidData[6]);
 
-      expect(res).to.have.status(422);
+      expect(res).to.have.status(400);
       expect(res.body).to.have.property('message').eql({ passwordLength: 'Password should be a minimum of 8 characters' });
     });
 
-    it('should return 422 if password is not alphanumeric', async () => {
+    it('should return 400 if password is not alphanumeric', async () => {
       const res = await chai.request(app)
         .post('/api/v1/auth/signup')
         .send(userMockData.invalidData[7]);
 
-      expect(res).to.have.status(422);
+      expect(res).to.have.status(400);
       expect(res.body).to.have.property('message').eql({ passwordType: 'Password should be alphanumeric' });
     });
 
-    it('should return 422 if required values not sent', async () => {
+    it('should return 400 if required values not sent', async () => {
       const res = await chai.request(app)
         .post('/api/v1/auth/signup')
         .send(userMockData.invalidData[8]);
 
-      expect(res).to.have.status(422);
+      expect(res).to.have.status(400);
       expect(res.body).to.have.property('message').eql([{ username: 'username field is required' }]);
     });
   });
