+++ swagger.yml
@@ -479,41 +479,6 @@ paths:
             status: 401
             error: Access Denied. Please Log In.
   
-  /article/comment/:commentid/history:
-    get:
-      summary: Handles getting comment edit history
-      description: Get all comments history by a commentId
-      tags:
-        - Comments
-      produces:
-        - application/json
-      security:
-        - Bearer: []
-      parameters:
-        - in: path
-          name: commentid
-          required: true
-          type: uuid
-          description: Id of the comment.
-      responses:
-        200:
-          description: Comment history was received successfully
-          example:
-            status: 200
-            message: Comment history retrieved successfully
-            data:
-              count: 2
-              rows:
-                - comment: "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nulla interdum | updated first"
-                  commentId: "c22c38cf-894c-417c-acf8-68eb0712bdaa"
-                  createdAt: "2019-04-16T15:02:34.602Z"
-                  updatedAt: "2019-04-16T15:02:34.602Z"
-        404:
-          description: Comment with the given id was not found in the database
-          example:
-            status: 404
-            error: Comment does not exist.
-  
   /article:
     post:
       tags:
@@ -976,101 +941,6 @@ paths:
                   email: "o.yes@gmail.com"
         503:
           description: "Some error occurred"
-  
-  /reports/categories:
-    get:
-      tags:
-      - Reports
-      summary: Get all report categories
-      description: This endpoint enables user to get all report categories
-      produces:
-        - application/json
-      security:
-       - Bearer: []
-      responses:
-        200:
-          description: "Successfully retrieved report categories"
-          example:
-            status: 200
-            message: "Report categories retrieved"
-            data:
-              count: 2
-              rows:
-                - id: "b2b67e1e-d40c-47ef-8abf-62e1a330d4ef"
-                  name: "Plagiarism"
-
-                - id: "b2b67e1e-d40c-47ef-8abf-62e1a330d4ef"
-                  name: "Plagiarism"
-                  
-              
-  /reports:
-    post:
-      tags:
-      - Reports
-      summary: "User can report an article"
-      description: "A new report to be created"
-      consumes:
-      - "application/json"
-      produces:
-      - "application/json"
-      security:
-       - Bearer: []
-      parameters:
-      - in: "body"
-        name: "body"
-        description: "User can report an article"
-        required: true
-        schema:
-          $ref: "#/definitions/Report"
-      responses:
-        201:
-          description: "Report submitted successfully"
-          example:
-            status: 201
-            message: "Report submitted successfully"
-            data:
-                comment: 'This article plagiarises another article that can be found here https://example.com/original-article'
-                typeId: "fc73ad52-1b04-4cba-a54f-e88b74555472"
-                articleId: "fc73ad52-1b04-4cba-a54f-e88b74555472"
-        400:
-          description: "Article does not exist"
-          example:
-            status: 400
-            message: "Article does not exist"
-        503:
-          description: "Some error occurred. Please try again"
-  /comment/:commentid/like:
-    post:
-      tags:
-      - Comments
-      summary: "Like a comment"
-      consumes:
-      - "application/json"
-      produces:
-      - "application/json"
-      security:
-       - Bearer: []
-      parameters:
-      - in: "path"
-        name: "commentid"
-        description: "A  signed in user can like a comment"
-        required: true
-      responses:
-        201:
-          description: "Comment liked successfully"
-          example:
-            status: 201
-            message: "Comment liked successfully"
-            data:
-              - like:
-                  commentid: "string"
-                  userid: "string"
-        409:
-          description: "You cannot like a comment twice"
-          example:
-            status: 409
-            message: "You cannot like a comment twice"
-
 
   /profiles/followers:
     get:
@@ -1339,15 +1209,3 @@ definitions:
     properties:
       articleId:
          type: "string"
-  Report:
-    type: 'object'
-    properties:
-      comment:
-        description: 'Comment associated with the report'
-        type: 'string'
-      typeId:
-        description: 'id of the corresponding Report type i.e. plagiarism, assualt, rules violation e.t.c'
-        type: 'string'
-      articleId:
-        description: 'id of the article being reported'
-        type: 'string'
